[tools/buildTools] Fixed some unclear errors output in rccd

+++ tools/buildTools/rccd.sh
@@ -1,8 +1,18 @@
 rccd()
 {
-    rccd_utils $@
+  if [ "$#" -ne 1 ];
+  then
+    echo "usage: rccd name"
+    return
+  fi
+  rccd_utils $@
+  if [ `ls -1 /tmp/*_rccd.output 2>/dev/null | wc -l` -gt 0 ];
+  then
     directory=$(cat /tmp/*_rccd.output | tr -cd '[:print:]' || echo "")
     rm /tmp/*_rccd.output > /dev/null 2>&1
     [ ! -z "$directory" ] && cd "$directory" && yaku || echo "No valid component found"
+  else
+    echo "usage: rccd name"
+  fi
 }
